Normalise in base: centralise code to join filepaths (#2771)
* normalise in path

* not using realpath; tests

* use new api in sync

* help

* in files

* use path.resolve instead of path.join

+++ packages/server/routes/sync.js
@@ -335,16 +335,15 @@ router.post(
   "/clean_sync_dir",
   error_catcher(async (req, res) => {
     const { dir_name } = req.body;
-    const safe_dir_name = File.normalise(dir_name);
     try {
       const rootFolder = await File.rootFolder();
-      const syncDir = path.join(
+      const syncDir = File.normalise_in_base(
         rootFolder.location,
         "mobile_app",
         "sync",
-        safe_dir_name
+        dir_name
       );
-      await fs.rm(syncDir, { recursive: true, force: true });
+      if (syncDir) await fs.rm(syncDir, { recursive: true, force: true });
       res.status(200).send("");
     } catch (error) {
       getState().log(2, `POST /sync/clean_sync_dir: '${error.message}'`);
